Fixed s3 plugin
S3's API doesn't like when you upload filenames with ':'s in them.
I suspect this has to do with the headers calculating a signature based
on your secret access key + the request performed, but then the ':' is
URL-encoded after the signature is calculated, and so it does not match
anymore. Solution: drop the ':'s.