Added support for purging archives
Every 30 minutes, a special tick comes through
to the run loop, indicating that archives should be
scanned for purgable archives, and purge operations requested.
These requests are sent to the agent defined in the supervisor's
PurgeAgent setting (the `purge_agent` in the config file).

If the purge tasks fail to execute on the agent, the records
remain untouched in the database, and purge will be attmepted again
until successful.

If a backup task fails, but returns a restore key, its archive is
created, and immediately marked as invalid, so that it can be purged

This change adds `status` and `purge_reason` to the archives table,
to enable differentiation between good and bad archives, as well as
troubleshoot why an archive was removed. Other miscellaneous schema
additions occurred for tracking archive data through tasks.

By default, the CLI will return only valid archives, but can be used
to query archives of varying status via the `--status` or `--all`
flags.

During the sweep/purge cycle, there is also a detection phase for
finding jobs whose expiry date has passed, so they can be picked
up on by the purging requests.