Fix Timeout Logic
We apparently were not setting TimeoutAt (leaving it at the zero value
of "a long time ago" which was triggering the timeout logic immediately
after a task was kicked off (and transitioned from schedq to runq).

In addition, cancellation of a timed out task was not removing the task
from the runq, which was causing us to loop repeatedly trying to cancel
the same tasks over and over again.

Fixes #72