Address redis backup failures during load
While under load, Redis may be appending to the
appendonly.aof file while we back it up. If this
happens, tar gets upset, unless we disable the warning.

Disabling this warning gives us a copy of the file that
tar saw at the beginning of the backup.  This does not fix
any potential corruption in the appendonly.aof file, merely
allows tar to complete successfully.

At restore time, the `redis-check-aof --fix` command is run
against each appendonly.aof file to correct any corruption.
The end result is that your restore may look like it was taken
1 second (the appendonly.aof sync interval) prior to when the
backup started.

Addresses #77