Allow custom error codes to be ok in plugins
Added support in the plugin framework to enable
plugin writers to specify what exit codes are
acceptable when running a command. In all
cases, an exit code of 0 is ok. `plugin.Exec()`
to only accepts exit codes of 0. To override,
you must use `plugin.ExecWithOptions()`, and
lose all the helpfulness that `plugin.Exec()`
provides to you.