Failed tasks/archives no longer report as successful
Due to confusing WorkerUpdate struct usage, failed tasks
were being `FAILED`, then subsequently `STOPPED`, to set
the time of job completion. `STOPPED` also sets status to `done`.
`FAILED` already sets the completion time to now. The `FAILED`
op has been updated to be more explicit about setting completion
time, and moved to the end of the worker runloop, so you either
have success, or failure.

Additionally, archives were not being invalidated, as the logic
for that would only be triggered if the creation of an archive
in the database failed. This has been addressed to happen
only if the archive database record was successfully created.

Fixes #79