Grab correct error codes/signals from SSH commands
The agent was improperly returning a hard-coded exit
status to the worker. We now grab the actual exit status
from the shield-pipe process that was run, or signal, if
it exited with a signal, and send that back to the worker,
formatted as big-endian, for the SSH server to decode.

Fixes #88