GetAnnotatedTask now works with purge/restore tasks
Moved GetAnnotatedTask to use db.TaskFilter when querying,
to get the same column mapping + basic query structure
as GetAllAnnotatedTasks, to future proof against changes
to the SQL query for retrieving tasks. This eliminated
an INNER + LEFT join that were not used anywhere, and
allowed records without job UUIDs to be returned.

Fixes #109