Enabled S3 plugin to talk to non-AWS endpoints
s3gof3r did not support AWS Signature Version 2,
which is required for many of the non-AWS S3 implementations.
Switched to minio-go, as it supports both, and is still pretty easy
to use.

Added support for choosing which signature version to use with the
S3 service. If not specified, defaults to version 4.

Also added SOCKS5 Proxy support for the s3 plugin, to ease testing of
various S3 endpoints.