Fix Weekly Timespec Bug
This commit addresses a rather common scheduling scenario with weekly
backups that will send SHIELD into a fit of constant job re-execution.
The Next() logic for weekly Spec objects doesn't properly advance to the
next week if the current day lands on the same weekday as the "next"
day.

In practice, this happens all the time, since the scheduler calculates
the Next() value immediately after the scheduled execution time (on the
same day).  The observed phenomenon is that the suervisor will keeping
seeing the "next" run time of a weekly job as "earlier today", until the
midnight boundary causes the buggy logic to forward to the next week.

This has been fixed, and regression tests added.