Enable UAA authentication/authorization support
Added support for the UAA to be used as an OAauth2 provider
for SHIELD, to handle authentication/authorization. If used,
you must specify the `auth.oauth.provider_url` configuration property,
to point at the UAA being used.

Authorization is performed by seeing if the authenticated user is
in a member of a UAA group(s) listed in `auth.oauth.authorization.orgs`.