Proper Weekly Re-scheduling
This commit fixes a bug in the weekly scheduling behavior of
timespec.Spec objects.  A weekly schedule for 3am UTC on Sunday would
kick off at exactly 03:00:00 UTC.  The supervisor would then reschedule
the job for 03:00:00 UTC (the bug). Then, the supervisor would tick
once, and at 03:00:01+0000 (1s later) the supervisor would kick off
another task for the backup job, and reschedule the job (again) for
03:00:00 UTC.

This would continue until 03:01:00 UTC, at which point the scheduling
logic would work past the bug, see that the original spec timestamp is
in the past, and forward it one week.

The upshot of this is that for weekly backups, you get 60 copies of your
backups.  For redundancy?

Fixes #159, by checking if the timespec represents the current time
(to 1 second of precision), or any time before it.  The check for
equality fixes the bug.