allow only a single task per target
Since task scheduling is not done in parallel, we can check prior to
starting a task whether it is allowed to run that task.

A task may not run if a task with same target uuid is running. This is
useful if a restore operation is currently in progress and another
backup should be initiated.