Bail out if no data is received by shield-pipe
Fixes #215

These modifications to shield-pipe correctly detect the case where a
plugin is not sending data to the other side.  For backup operations,
this catches bad backups.

The `tail -c1` command grabs up to the last byte out of the stream and
writes it to a temporary file.  This should help us to keep the disk
impact as low as possible, but it does mean that we know rely on disk
to be writable (at least `/tmp`).  Note that we cannot use `head -c1`
here, because of the `set -o pipefail`, the whole pipeline will SIGPIPE
right after `head -c1` reads the first byte, and then exits
("prematurely")