Add support for mysqldump options in mysql plugin
This commit adds support for mysqldump options in the mysql plugin.
The setting "mysql_options" allows setting common mysqldump options
like --force and/or --single-transaction.

This gives the MySQL Admins more power on how shield runs the
backup.

If the database contains VIEWS which are invalid, mysqldump will
exit with an error and the backup will fail. Adding --force will
make mysqldump ignore the error and create a proper backup (altough
the DB is probably not in the best state, but things usually won’t
get better when backuping ... )