Add functionality to ignore the ExitStatus
This commit introduces CheckExitStatus (default: true) for plugin
execs. If CheckExitStatus is set to false, the process will ignore
the exit status and will set the action to sucess even when the
invoked backup mechanism exited with an exit code != 0.

The support is currently only built into the mysql plugin.

Unfortunately this is implemented using the strings "true" or
"false" because of my lacking go skills.

But it works. I have tested it using a DEV shield-boshdeployment
release.