Job Status Enhancements (#271)
Re-introduce the `/v1/status/jobs` endpoint for showing the most recent success or failure of a job task.  Intended for a monitoring/alerting application to scrape this information periodically.  The UI also reflects this information, under Jobs the status is now color coded.  Also on the dashboard recently failed jobs will appear in a notification bar at the top.