Core Rewrite
This massive commit introduces several interrelated and far-reaching
changes to the SHIELD codebase, in line with the 2017-18 ROADMAP.

The supervisor core has been completely rewritten as the `core` module,
and it depends more heavily on orchestration via the SHIELD database
than the resync'd supervisor implementation ever did.

The SHIELD agent has been refreshed with some more idiomatic Go
patterns, as well as agent auto-registration and health / capability
monitoring support (via the "status" op).

The Web UI is getting a massive overhaul, with a new design and a new
approach to UX that tries to antipicate common operator tasks and make
them as simple and straightforward as possible.

This is still a work-in-progress as part of the v7 rewrite of SHIELD.
Notably missing from this revision is API authentication; it was removed
as part of the rewrite and will be re-integrated at a later time, in a
future commit.

If you are reading this message because of a `git blame`, you have my
deepest and sincerest apologies; it is not my intent to break existing
functionality, only to improve the product.