Added -k support for create-backend
* If -k is given during backend creation, that backend will always skip
SSL validation when used.
* A new section, properties, was added to the configuration file,
which will store all future added metadata for backends.
skip_ssl_validation info is stored here.
* `backends` now shows (insecure) after the name if the backend is
marked to have SSL skipped
* `backends` now displays the currently targeted backend in green
* Because of the growing complexity of the config file, an abstraction
layer was put on top of the config layout to allow the api (and cli) to
have a cleaner interface through which to access and provide backend
info. It can be found at api.Backend. The currently targeted backend is
loaded into this format when the config is loaded.
* Config was moved into a separate package under CLI, because the config
file is not a general API thing.
* The config file is now saved whenever shield quits and the run command
was successful, instead of being the responsibility of the API calls or
specific commands (looking to optimize to only save if a change has
actually occurred).