SHIELD Core Init and Unlock UI
This commit introduces new UI screens for initializing a new SHIELD core
and unlocking a locked SHIELD core, complete with the API changes that
needed to accompany.

`/init.js` is a new Core endpoint, which serves up the health + auth
information of `/auth/id` and `/v2/health` in one nice,
browser-parseable package with the added bonus of blocking the UA
rendering until we get the data we need to actually do the render.  This
drops out the `async: false` warning that Mozilla browsers were
emitting, and makes things smoother for the end user.