Schedules removed from CLI
* Schedule commands and help group removed outright.
* Job displays now show the timespec that defines a schedule, regardless
of backend API version
* Different code paths in jobs for v1 and v2 API.
* TUI forms now have a function to lookup a field by name, in order to
determine what the user input was.
* ShowX functions were moved out of the internal package, and moved into
each X package (e.g. ShowJob was moved into the jobs package and renamed
to Show)
* For v1, a new schedule is created with the given timespec when a job
is created.
* For v1, when a job is edited or deleted, a reference count is taken
for the old schedule and if the reference count is zero, a request to
delete the schedule is made.
* For v2, the timespec is just given as the schedule field, as per the
new spec. Easy peasy.