CLI added login, logout, and whoami
* Auth prompting was taken out of the api package and moved into CLI
commands. A user now needs to authenticate manually with one of these
commands, as opposed to the old way, in which the user would simply be
prompted upon receiving a 401
* API functions were added for contacting POST /v2/auth/login, GET
/v2/auth/logout, and GET /v2/auth/id.
* Parsing functions for getting v1 and v2 API error responses were
added. v1 endpoints on v2 APIs will currently be unable to get error
messages, but this will be okay once the v1 endpoints go away.
* Error types were added for 401s and 403s, namely ErrUnauthorized and
ErrForbidden. These can be type checked to determine if an api error was
due to authentication problems.
* The core API exposes test endpoints if the debug flag in the daemon
config is set.
* The v2 api now gives 401s if the user is not authenticated (as opposed
to 403)
* The api.Backend type now accepts an APIVersion int, for being told
which APIVersion to use. This value can be stored in the in-memory
version of the config for use in the rest of the CLI, but is never written out
to disk.
* testdev now sets the debug flag for the shield daemon config.
* The login endpoint now gives back information about the user which was
authenticated in the same format as /auth/id
* api_version is no longer exposed by the status endpoint (because you
need to be authenticated to reach it), and is instead exposed by
/v1/ping