New UI flow for CLI copy-paste of creds
The new `#!/cliauth` page now prints out the session ID, which it
expects to receive via the `s:...` hash path parameter.  The
authentication provider handlers are properly redirecting to this for
CLI flows.

This necessitated extracting all of the session handling out of the
individual authentication providers, and moving it into the core handler
so that we could disposition based on web/cli flows.