Validate Authentication Providers and pre-create Tenants
Now, if the configuration for an authenticaiton provider is bogus, we
learn about it early, before we even bind the API listener, so that
someone can go fix it.  We also pre-create the referenced tenants, so
that admins aren't "surprised" to find new and/or missing tenants when
they look at the admin backend.  Auto-vivification is still there.