Rework Token Authentication
The Token Authentication Provider is no more.  Instead, SHIELD now handles
token authentication exceptionally, as it does local authentication.

The X-Shield-Token header can now be set, as it was in pre-v8 versions of
SHIELD, to authenticate via an auth token.  These auth tokens are generated
by an authenticated user, and amount to little more than "extra" sessions tied
to that users account.

A new `/v2/auth/tokens/*` endpoint has been added to allow for provisioning,
viewing, and revocation of auth tokens.  The CLI has been updated to include
new commands for doing likewise.

Since this removes the ability to do "initialization" via a token, the SHIELD Core
now honors a _failsafe_ user, defined in the SHIELD Core configuration YAML
file.  This failsafe username / password will be added to the local database on
boot, if (and only if) there are no other administrative users.