Configuration Wizard
The #!/do/configure web UI page now walks the user through a
configuration wizard for setting up a new job, optionally allowing them
to create targets, stores, and retention policies along the way.

The "create" operations in the `/v2` endpoints for targets, stores, and
retention policies, picked up a new `?test=t` query string parameter for
running in _test mode_, where the actual thing will not be created, but
the validation will still be performed.  The Web UI needs this.