Updates to bin/testdev / build process
Plugins are now built from a central manifest file (`/plugins`), and all
plugin-related make targets rely on this.  Moving forward, adding new
plugins to the build system is as simple as appending a line to that
file.

The testdev script now properly rebuilds all of its dependencies, so you
can do this:

  make clean dev

And it just works.