Fix js handler stacking bugs
By attaching handles to the return value of `$('#main').html(...)`, we
end up double- (or triple-, or ntuple-) firing events for things like
'submit form'.  This causes no end of zaniness and confusion, so
instead, we now attach event handlers to the argument to `html()`, the
template root DOM element.