Improve `buckler import`
It no longer requires (or uses) the --core / $SHIELD_CORE configuration
stuff; and instead relies _exclusively_ on the core / token
configuration in the import file.  This fixes #384.

A new top-level `insecure_skip_verify` configuration option in the
import yaml allows you to skip the TLS verification (in case the cert
has expired, or is for a different IP)

YAML, not JSON.