Add the ability to restore from previous SHIELD on initialization
When an operator starts a fresh SHIELD instance (in the case of
catastrophic failure, etc.) a new initialization screen is displayed
that allows the upload of a previous SHIELD self-backup. An out-of-band
restore task is fired, and then shield returns to the login screen.

A few caveats to note: this screen replaces the default initialization
page. This means, in the case of Vault being blown away, but SHIELD has
not been touched, the user is still given the option to restore from a
previous SHIELD state (though they can simply just enter a new master
password in and be given a fixed key as per previous logic)

This task assumes that SHIELD data is located in
`/var/vcap/store/shield/`. It was assumed the use-case of SHIELD would
be in a BOSH release.

The ability for the FS plugin to "retrieve" files has been re-added to
support this feature. All other "store" functionality stills remains
unimplemented as FS remains a plugin unintended to store data.