Create ShieldRestoreOperation, and make SRO more resilient
After reviewing the previous self-bootstrapping restore functionality
from `413e508` and `493a0ee`, we believe it was important to further
flesh out bootstrapping.

This commit brings ShieldRestoreOperation, specifically for use in
out-of-band restores that directly call an agent. This allows us to
receive logging output for the task, and in the case of SRO failure,
debug the problem. The WebUI has also been updated to automatically
display the log if it detects a failed SRO. Numerous minor changes were
made to support a "full" SRO, including the introduction of a core flag
that allows for shutting down SHIELD (`core.seppuku`)

To facilitate an easier restore process, we've removed the `DBPath` and
`VaultKeyfile` configuration options and added a `DataDir`, which stores
SHIELD-related data, as well as the Vault directory.

Also included within this commit are measures to make
`/v2/bootstrap/restore` more resilient to user input errors, including
malformed fixed keys, and not panicing on simple errors. SRO now also
properly stores the bootstrap restore in the OS-defined temporary
directory, rather than a hardcoded `/tmp/`