Support SHIELD Initial-Restore
When an operator starts a fresh SHIELD instance (in the case of
catastrophic failure, etc.) a new initialization screen is displayed
that allows the upload of a previous SHIELD self-backup. An out-of-band
restore task is fired, and then SHIELD returns to the login screen.

The ability for the `fs` plugin to "retrieve" files has been re-added to
support this feature. All other "store" functionality stills remains
unimplemented and `fs` remains unable to act as a store plugin.

SHIELD now gets configured with a data directory, instead of paths to
the database and vault locations.  It also picked up a new API endpoint,
`/v2/bootstrap/restore` which enables the initialization screen to also
show failures from a previous init-restore.