Move SHIELD encryption to a standalone binary (#400)
After manual testing and debugging, we've discovered that SHIELD was not
encrypting the backups it was taking. This was due to two issues:
`shield-pipe` was not properly reading the appropriate
encryption-related environment variables, but also the change in plugin
`Exec()` structure allowed some plugins, such as `fs`, to skip
encryption wholesale.

This commit solves both problems, by fixing `shield-pipe` to properly
test environment variables, and by removing the
encryption/decryption-related code from `Exec()` and into a standalone
`shield-crypt` binary that takes the compressed backup from a pipe to
`--encrypt` or `--decrypt`.