Add "Errored" state when `/v2/health` throws an error
Currently, when the backend is unable to get the status of certain
services (e.g. Vault), the endpoint throws a 500 and the web UI fails to
render (it's expected a proper health response).

This is an attempt to fix this, by introducing a hard-coded "ERRORED"
state on a 500 response. It sets all health values to false and empty
data storage arrays. It's designed to inform the operator that SHIELD is
not in a good state, but avoids being overly verbose.