Prevent Partial Completion of Wizard Steps
The do-configure wizard now requires that users select a system before
moving on to the storage step, that they pick a store before moving to
job details step, and that they fill out all required job detail form
fields before moving on to review.

The do-backup wizard now requires that users select a system before
moving on to the job selection step, and that they pick a job (store)
before moving on to review.

The do-restore wizard now requires that users select a system before
moving on to the archive selection step, and that they pick an archive
before moving on to review.