Strict Mode for fs plugin
The new `strict` setting in the filesystem plugin allows the operator to
decide if files that turn up missing during the `stat()` call of the
base path walker function should trigger failure of the overall backup
("strict":"yes") or if they should be ignored ("strict":"false").

The default is "false", as that is what the majority of people expect.

The `fs` plugin now also sports some useful normal-fare verbose output,
to give some indication to operators of what is going on while tailing
the task output in the web UI, or the CLI.