Re-work Documentation
Documentation now lives under either `docs/ops` for operations-facing
documents, or `docs/dev` for stuff that mostly pretains to SHIELD
developers and plugin authors.

These documents are maintained in markdown, are location-agnostic, and
can be built with `bin/mkdocs`, to be embedded in the shieldproject.io
website.  See `docs/README.md` for the full story.

As part of this reorg, some old documentation (which was mostly
incomplete) has been lost.  Some historical documents were maintained,
for nostalgia.  Everything is being rewritten.