Supply purge tasks with their restore keys
Purge tasks are now being properly supplied with the restore key
necessary for deleting the archive blob.  Fixes #516.

We now also have a new fixups system in the database layer, to allow
the SHIELD core to automatically apply data fixes that are not
_technically_ schema migrations, but do clean up from past bugs.

The first of these is purge-tasks-516, which undoes the damage done by
the fixed bug in the wild, by marking all archives that failed to purge
as "expired", forcing a retry by the SHIELD core.