Handle AWS 301's without Location: headers
A 301 response without a Location header is an error per the HTTP RFCs.
But Amazon returns a 301 without a Location header precisely for what it
calls the PermanentRedirect error code.  This is used, during path-based
addressing to point to a bucket outside of the US regions.

This commit causes the s3 plugin to re-attempt actions when the returned
response violates this part of the HTTP speec, and re-attempt without
path-based addressing.

Fixes #568