Fix Websocket timeouts (#611)
The server now sends keepalive messages after a configurable timeout
interval if a message has not been sent during that time. This should
keep the connection open even if there are no events to send.

The write timeout for messages is now configurable.

The configuration will now error if mbus parameters are negative.