Environmental Configuration of SHIELD
This changeset allows the SHIELD Core (shieldd) and the SHIELD Agent
(shield-agent) to be configured completely (minus auth providers) via
environment variables.  This is all but required for running SHIELD
effectively in Docker, Docker Compose, and Kubernetes.

To do so, I also had to incorporate changes requested in #618, to remove
the variable-length `fabrics:` configuration.

`bin/testdev` has been updated to incorporate an environment-first
approach, as a proof-of-concept.

Fixes #618.