Handle Tenant Invite / Banishment in Web UI (better)
Now, as administrators (via either the CLI or the Web UI) add
people to tenants, and remove them from tenants, the websocket
and the top bar navigation menu update in real time.

If you are added to a tenant (or your role is changed) the account
menu will adjust accordingly.  When you get added to your first
tenant, it will become your active tenant.

When you are removed from a tenant, it will disappear from the
account menu, and if it was your active tenant, a new active tenant
will be selected.  If i was your last tenant, you will go back to having
"no tenants" and nothing useful to do in the SHIELD Web UI.