Reverse Global Tenant check logic in cancel task endpoint (#659)
Attempting to cancel a task by DELETEing at the tenant specific task API
endpoint would result in No Such Task being returned. It appears that
this is because of a reversed logic check, which prevents this endpoint
from being a workaround to delete global tasks. Instead, it ONLY allows
the cancellation of global tasks.