Stop leaking sql prepared statements.
Previously, we were caching prepared statements, so the amount of
statements we actually compiled was bounded, or at least grew far
slower. Now we make a new statement for every request, but we weren't
closing the statements, causing the sql database layer to never release
the statement dependency resources. This causes the memory use of SHIELD
to grow unbounded. Now, the statements are closed when no longer needed,
and the memory released, and the world saved.