Externalize Database to PostgreSQL
We are no longer on SQLite, and instead are targeting
PostgreSQL natively.  Since Postgres can run in a container,
and be accessible only on loopback, or on a c2c network,
this is just as good as SQLite, with better centralization
prospects.

We will not, at this time, be entertaining _other_ external
RDBMs provides (no MySQL, no Oracle, no SQL Server).