Fix for infinite loop for misbehaving WebSocket proxies
When shield is fronted by a reverse proxy which hides the 401 authentication status to the websocket handshake and rather returns a `HTTP/1.1 101 Switching Protocols` response, then shield client code enters an infinite loop of requests to `v2/events` and `v2/bearings`.

This comes from the fact that the error handling code to the 401 response status (which redirects the browser to the homepage) does not trigger. Instead, the websocket open event triggers a `v2/bearings` ajax call whose handler fails to parse the following unauthorized response with the following trace

```json
{"vault":"","shield":{"api":2,"version":"8.7.2","env":"sandbox","color":"yellow","motd":"Welcome to SHIELD!\n"},"user":null,"stores":null,"tenants":null}
```

```
data.js:489 Uncaught TypeError: Cannot read property 'length' of null
    at Object.success (data.js:489)
    at i (jquery.js:2)
    at Object.fireWith [as resolveWith] (jquery.js:2)
    at A (jquery.js:4)
    at XMLHttpRequest.<anonymous> (jquery.js:4)
```

As a result of this unhandled error, the websocket is closed. When the `v2/events` websocket closes, the shield client immediately reopens a new websocket, entering an infinite loop of requests to `v2/events` and `v2/bearings`

This infinite loop exhausts client and server resources. As a result, when the login form is submitted, the login ajax requests seems to loose the race with other requests or possibly be cancelled by the shield error handling, resulting in the login page to never succeed.

This commit adds some delays (3s) in the websocket close event handler as to slow down this infinite loop. It improves the situation much, whilea stalled login forms was still observed once out of 20 successful logins.